Verify hue light off with 5s interval
Noticed that once in a while the hue bridge seem to ignore the off
command. Therefore lets get the light state json string and search for
on="true" as an indication the light is still on. Note that if the http
response come in chunks, we may miss this sub-string.
Lets see whether this really works, this is hard to reproduce. A simple
test to turn the light back on using the Hue mobile app within this 5s
time frame, properly turned the light back off.